Load dataset file by filename, outsource utilities
